# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 400
        example: 400
        type: integer
      details:
        description: Field-level validation error details (for validation errors)
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      error:
        description: |-
          Error type/category
          example: Bad Request
        example: Bad Request
        type: string
      message:
        description: |-
          Detailed error message
          example: Invalid email format
        example: Invalid email format
        type: string
      request_id:
        description: |-
          Request ID for tracing and debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  models.FieldError:
    properties:
      code:
        description: |-
          Error code for programmatic handling
          example: email
        type: string
      field:
        description: |-
          The field name that failed validation
          example: email
        type: string
      message:
        description: |-
          Human-readable error message
          example: Invalid email format
        type: string
      value:
        description: The rejected value (excluded for sensitive fields)
    type: object
...
  models.UsageEventRequest:
    properties:
      event_type:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      quantity:
        type: integer
      resource:
        type: string
      unit:
        type: string
    required:
    - event_type
    - resource
    type: object
  models.UsageLimits:
    properties:
      api_calls:
        type: integer
      compute_ms:
        type: integer
      file_uploads:
        type: integer
      storage_bytes:
        type: integer
    type: object
  models.UsagePercentages:
    properties:
      api_calls:
        type: integer
      compute_ms:
        type: integer
      file_uploads:
        type: integer
      storage_bytes:
        type: integer
    type: object
  models.UsageSummaryResponse:
    properties:
      limits:
        $ref: '#/definitions/models.UsageLimits'
      limits_exceeded:
        type: boolean
      percentages:
        $ref: '#/definitions/models.UsagePercentages'
      period_end:
        type: string
      period_start:
        type: string
      totals:
        $ref: '#/definitions/models.UsageTotals'
    type: object
  models.UsageTotals:
    properties:
      api_calls:
        type: integer
      compute_ms:
        type: integer
      file_uploads:
        type: integer
      storage_bytes:
        type: integer
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/react-golang-starter-kit/issues
  description: A comprehensive REST API for the React Go Starter Kit application built
    with Fiber, GORM, and PostgreSQL. This API provides secure user authentication,
    user management, and system health monitoring.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/react-golang-starter-kit
  title: React Go Starter Kit API
  version: 1.0.0
paths:
...
  /api/usage:
    get:
      description: Returns usage metrics for the current billing period
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsageSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get current usage summary
      tags:
      - Usage
  /api/usage/alerts:
    get:
      description: Returns unacknowledged usage alerts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get usage alerts
      tags:
      - Usage
  /api/usage/alerts/{id}/acknowledge:
    post:
      consumes:
      - application/json
      description: Marks a usage alert as acknowledged
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Acknowledge usage alert
      tags:
      - Usage
  /api/usage/history:
    get:
      description: Returns usage metrics for past billing periods
      parameters:
      - description: Number of months to retrieve (default 6)
        in: query
        name: months
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UsageSummaryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get usage history
      tags:
      - Usage
  /api/usage/record:
    post:
      consumes:
      - application/json
      description: Records a usage event for metering purposes
      parameters:
      - description: Usage event details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UsageEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Record usage event
      tags:
      - Usage
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Format: `Authorization:
      Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"